custom sekizai processor to separate compressable css and js files
