+++ example/myshop/templates/myshop/pages/base.html
@@ -10,7 +10,7 @@
 	<meta name="description" content="{% block meta-description %}{% endblock %}" />
 	{% block head %}{% endblock head %}
 	{% render_block "ext-css" %}
-	{% render_block "css" postprocessor "myshop.sekizai_processors.compress" %}
+	{% render_block "css" postprocessor "shop.sekizai_processors.compress" %}
 </head>
 
 {% addtoblock "js" %}<script src="{% static 'node_modules/picturefill/dist/picturefill.min.js' %}" type="text/javascript"></script>{% endaddtoblock %}
@@ -45,7 +45,7 @@
 	</footer>
 
 	{% render_block "ext-js" %}
-	{% render_block "js" postprocessor "myshop.sekizai_processors.compress" %}
+	{% render_block "js" postprocessor "shop.sekizai_processors.compress" %}
 	<script type="text/javascript">
 	angular.module('myShop', ['ngAnimate', 'ngSanitize',
 		{% render_block "ng-requires" postprocessor "djng.sekizai_processors.module_list" %}
